test: add for style injection
