--- src/ui/useTheme/ThemeProvider.test.tsx
@@ -1,1 +1,0 @@
-it('does something', () => {});
