+++ src/ui/useTheme/index.ts
@@ -1,6 +1,2 @@
-export { default } from './ThemeProvider';
-export type { ThemeProps } from './ThemeProvider';
-
-export { default as isThemeDark } from './isThemeDark';
-
-export { default as useTheme } from './useTheme';
+export { default, isThemeDark, useTheme } from './useTheme';
+export type { ThemeProps } from './useTheme';
