+++ src/ui/useTheme/useTheme.test.tsx
@@ -1,8 +1,7 @@
 import { act, renderHook } from '@testing-library/react-hooks';
 import { mockMatchMedia } from 'helpers/test';
 
-import ThemeProvider from './ThemeProvider';
-import useTheme from './useTheme';
+import ThemeProvider, { STYLE_RULES, isThemeDark, useTheme } from './useTheme';
 
 function renderTheme() {
   return renderHook(() => useTheme(), {
@@ -10,40 +9,61 @@ function renderTheme() {
   });
 }
 
-it('returns base mode and device types', () => {
-  mockMatchMedia({ matches: false });
+describe('ThemeProvider', () => {
+  jest.useFakeTimers();
 
-  const { result } = renderTheme();
+  it('returns base mode and device types', () => {
+    mockMatchMedia({ matches: false });
 
-  expect(document.documentElement.getAttribute('data-theme')).toBe('light');
-  expect(document.documentElement).toHaveClass('theme--light');
-  expect(result.current.device.mode).toBe('light');
-  expect(result.current.device.isTablet).toBe(false);
-  expect(result.current.device.isDesktop).toBe(false);
-  expect(result.current.device.isTv).toBe(false);
+    const { result } = renderTheme();
+
+    expect(document.documentElement.getAttribute('data-theme')).toBe('light');
+    expect(document.documentElement).toHaveClass('theme--light');
+    expect(result.current.device.mode).toBe('light');
+    expect(result.current.device.isTablet).toBe(false);
+    expect(result.current.device.isDesktop).toBe(false);
+    expect(result.current.device.isTv).toBe(false);
+  });
+  it('returns true for device types and dark color scheme if media matches', () => {
+    mockMatchMedia({ matches: true });
+
+    const { result } = renderTheme();
+
+    expect(document.documentElement.getAttribute('data-theme')).toBe('dark');
+    expect(document.documentElement).toHaveClass('theme--dark');
+    expect(result.current.device.mode).toBe('dark');
+    expect(result.current.device.isTablet).toBe(true);
+    expect(result.current.device.isDesktop).toBe(true);
+    expect(result.current.device.isTv).toBe(true);
+  });
+  it('returns user prefered color scheme instead of system one', () => {
+    mockMatchMedia({ matches: false });
+
+    const { result } = renderTheme();
+
+    expect(document.documentElement.getAttribute('data-theme')).toBe('light');
+    expect(document.documentElement).toHaveClass('theme--light');
+    expect(result.current.device.mode).toBe('light');
+    act(() => result.current.device.setMode('dark'));
+    expect(document.documentElement.getAttribute('data-theme')).toBe('dark');
+    expect(document.documentElement).toHaveClass('theme--dark');
+    expect(result.current.device.mode).toBe('dark');
+  });
+  it('injects transition style rules on color scheme change', () => {
+    mockMatchMedia({ matches: true });
+    const { result } = renderTheme();
+
+    act(() => result.current.device.setMode('light'));
+    expect(document.head.children[0].textContent).toMatch(STYLE_RULES);
+    jest.runAllTimers();
+    expect(document.head.children[0]).toMatch('');
+  });
 });
-it('returns true for device types and dark color scheme if media matches', () => {
-  mockMatchMedia({ matches: true });
-
-  const { result } = renderTheme();
-
-  expect(document.documentElement.getAttribute('data-theme')).toBe('dark');
-  expect(document.documentElement).toHaveClass('theme--dark');
-  expect(result.current.device.mode).toBe('dark');
-  expect(result.current.device.isTablet).toBe(true);
-  expect(result.current.device.isDesktop).toBe(true);
-  expect(result.current.device.isTv).toBe(true);
-});
-it('returns user prefered color scheme instead of system one', () => {
-  mockMatchMedia({ matches: false });
-
-  const { result } = renderTheme();
-
-  expect(document.documentElement.getAttribute('data-theme')).toBe('light');
-  expect(document.documentElement).toHaveClass('theme--light');
-  expect(result.current.device.mode).toBe('light');
-  act(() => result.current.device.setMode('dark'));
-  expect(document.documentElement.getAttribute('data-theme')).toBe('dark');
-  expect(document.documentElement).toHaveClass('theme--dark');
-  expect(result.current.device.mode).toBe('dark');
+describe('isThemeDark', () => {
+  it('returns true if scheme color is dark', () => {
+    expect(isThemeDark('dark')).toBe(true);
+  });
+  it('returns false if scheme color is light', () => {
+    expect(isThemeDark('light')).toBe(false);
+  });
 });
